fix: address code review issues — filtered index columns, float division, backoff cap
Agent-Logs-Url: https://github.com/ivanopaulon/EventForge/sessions/4d185e2a-aff0-4459-9f94-ec5b2319ae19

+++ EventForge.Server/Services/Monitoring/MonitoringService.cs
@@ -78,7 +78,7 @@ public class MonitoringService : IMonitoringService
                     && p.TenantId == tenantId.Value
                     && p.MaxUses.HasValue
                     && p.MaxUses.Value > 0
-                    && p.CurrentUses * 100 / p.MaxUses.Value >= 80)
+                    && (double)p.CurrentUses * 100 / p.MaxUses.Value >= 80)
                 .OrderByDescending(p => p.CurrentUses)
                 .Take(topN)
                 .Select(p => new PromotionUsageItemDto
